Refactor, fix bugs, and reword comments in initial_library.fs
Some of the comments seemed a bit off. Additionally, there was a bug in AGAIN
in which it didn't actually compile a jump. Also, WHILE's compile-time stack
effect is now consistent with gforth, which means we can have multi-exit while
loops. Also, added UNTIL. Also, added [CHAR] so we don't have to keep looking up
ascii values for characters. Also, whichever bug it was that caused problems
with splitting definitions up across multiple lines is now gone, so long live
the newlines!